refactor: remove unused imports, commented code and rename singlePlayerHandler
